+++ public/app/leaflet-extensions/MageControls.js
@@ -7,7 +7,7 @@ L.Control.MageFeature = L.Control.extend({
   onAdd: function () {
     var container = L.DomUtil.create('div', 'leaflet-bar leaflet-control');
     this._link = L.DomUtil.create('a', '', container);
-    this._icon = L.DomUtil.create('i', 'fa fa-map-marker icon-sage', this._link);
+    this._icon = L.DomUtil.create('i', 'fa fa-map-marker icon-mage', this._link);
 
     this._link.href = '#';
     this._link.title = 'New Observation';
@@ -37,7 +37,7 @@ L.Control.MageUserLocation = L.Control.extend({
     var container = L.DomUtil.create('div', 'leaflet-bar leaflet-control');
 
     this._locateLink = L.DomUtil.create('a', '', container);
-    this._locateIcon = L.DomUtil.create('i', 'fa fa-location-arrow icon-sage', this._locateLink);
+    this._locateIcon = L.DomUtil.create('i', 'fa fa-location-arrow icon-mage', this._locateLink);
     this._locateLink.href = '#';
     this._locateLink.title = 'Locate Me';
     this._locate = false;
@@ -50,7 +50,7 @@ L.Control.MageUserLocation = L.Control.extend({
       .on(this._locateLink, 'click', this._onLocateClick, this);
 
     this._broadcastLink = L.DomUtil.create('a', '', container);
-    this._broadcastIcon = L.DomUtil.create('i', 'icon-megaphone icon-sage', this._broadcastLink);
+    this._broadcastIcon = L.DomUtil.create('i', 'icon-megaphone icon-mage', this._broadcastLink);
     this._broadcastLink.href = '#';
     this._broadcastLink.title = 'Report Location';
     this._broadcast = false;
@@ -71,16 +71,16 @@ L.Control.MageUserLocation = L.Control.extend({
     this.startLocate();
 
     L.DomUtil.addClass(this._broadcastLink, 'leaflet-control-inverse');
-    L.DomUtil.addClass(this._broadcastIcon, 'icon-sage-inverse');
-    L.DomUtil.removeClass(this._broadcastIcon, 'icon-sage');
+    L.DomUtil.addClass(this._broadcastIcon, 'icon-mage-inverse');
+    L.DomUtil.removeClass(this._broadcastIcon, 'icon-mage');
   },
 
   stopBroadcast: function() {
     if (!this._broadcast) return; // broadcast already off
 
     L.DomUtil.removeClass(this._broadcastLink, 'leaflet-control-inverse');
-    L.DomUtil.addClass(this._broadcastIcon, 'icon-sage');
-    L.DomUtil.removeClass(this._broadcastIcon, 'icon-sage-inverse');
+    L.DomUtil.addClass(this._broadcastIcon, 'icon-mage');
+    L.DomUtil.removeClass(this._broadcastIcon, 'icon-mage-inverse');
 
     this._broadcast = false;
   },
@@ -96,8 +96,8 @@ L.Control.MageUserLocation = L.Control.extend({
     });
 
     L.DomUtil.addClass(this._locateLink, 'leaflet-control-inverse');
-    L.DomUtil.addClass(this._locateIcon, 'icon-sage-inverse');
-    L.DomUtil.removeClass(this._locateIcon, 'icon-sage');
+    L.DomUtil.addClass(this._locateIcon, 'icon-mage-inverse');
+    L.DomUtil.removeClass(this._locateIcon, 'icon-mage');
   },
 
   stopLocate: function() {
@@ -113,8 +113,8 @@ L.Control.MageUserLocation = L.Control.extend({
     }
 
     L.DomUtil.removeClass(this._locateLink, 'leaflet-control-inverse');
-    L.DomUtil.addClass(this._locateIcon, 'icon-sage');
-    L.DomUtil.removeClass(this._locateIcon, 'icon-sage-inverse');
+    L.DomUtil.addClass(this._locateIcon, 'icon-mage');
+    L.DomUtil.removeClass(this._locateIcon, 'icon-mage-inverse');
 
     this.stopBroadcast();
   },
@@ -149,10 +149,6 @@ L.Control.MageUserLocation = L.Control.extend({
       }
     });
   }
-	// _broadcast: function() {
-	// 	if (!this._location || !this.options.onBroadcast) return;
-	// 	this.options.onBroadcast(this._location);
-	// }
 });
 
 L.Control.MageListTools = L.Control.extend({
@@ -166,7 +162,7 @@ L.Control.MageListTools = L.Control.extend({
     var linkClass = this.options.enabled ? 'leaflet-control-inverse' : '';
     this._link = L.DomUtil.create('a', linkClass, container);
 
-    var iconClass = this.options.enabled ? 'fa fa-bars icon-sage-inverse' : 'fa fa-bars icon-sage';
+    var iconClass = this.options.enabled ? 'fa fa-bars icon-mage-inverse' : 'fa fa-bars icon-mage';
     this._icon = L.DomUtil.create('i', iconClass, this._link);
 
     this._link.href = '#';
@@ -184,16 +180,35 @@ L.Control.MageListTools = L.Control.extend({
     return container;
   },
 
+  hideFeed: function(hide) {
+    var hidden = L.DomUtil.hasClass(this._icon, 'icon-mage');
+    if (hidden === hide) return;
+
+    if (hidden) {
+      L.DomUtil.addClass(this._link, 'leaflet-control-inverse');
+      L.DomUtil.addClass(this._icon, 'icon-mage-inverse');
+      L.DomUtil.removeClass(this._icon, 'icon-mage');
+    } else {
+      L.DomUtil.removeClass(this._link, 'leaflet-control-inverse');
+      L.DomUtil.addClass(this._icon, 'icon-mage');
+      L.DomUtil.removeClass(this._icon, 'icon-mage-inverse');
+    }
+  },
+
   _toggle: function () {
-    var on = L.DomUtil.hasClass(this._icon, 'icon-sage-inverse');
+    var on = L.DomUtil.hasClass(this._icon, 'icon-mage-inverse');
     if (on) {
       L.DomUtil.removeClass(this._link, 'leaflet-control-inverse');
-      L.DomUtil.addClass(this._icon, 'icon-sage');
-      L.DomUtil.removeClass(this._icon, 'icon-sage-inverse');
+      L.DomUtil.addClass(this._icon, 'icon-mage');
+      L.DomUtil.removeClass(this._icon, 'icon-mage-inverse');
     } else {
       L.DomUtil.addClass(this._link, 'leaflet-control-inverse');
-      L.DomUtil.addClass(this._icon, 'icon-sage-inverse');
-      L.DomUtil.removeClass(this._icon, 'icon-sage');
+      L.DomUtil.addClass(this._icon, 'icon-mage-inverse');
+      L.DomUtil.removeClass(this._icon, 'icon-mage');
+    }
+
+    if (this.options.onToggle) {
+      this.options.onToggle(!on);
     }
   }
 });