+++ CHANGELOG.md
@@ -3,10 +3,11 @@ All notable changes to this project will be documented in this file.
 Adheres to [Semantic Versioning](http://semver.org/).
 
 ---
-## 4.1.2 (TBD)
+## 4.2.0 (TBD)
 
 * TBD
 ##### Features
+* New numberfield in event form
 * Rework observation/people feed to show map and feed on smaller devices
 * Allow user to upload avatar when creating an account